gen2 auth rework
